u-boot-marvell: Rework recipe
The recipe now uses the vendor name instead of the SoC family; this
allow for the same recipe to be used across different SoC families and
avoids code duplication.

Another import change is that we now use an unified repository for
U-Boot for all Marvell's SoCs. This also removes the indirection for
the source as this will use just public code.

Change-Id: I8a002fa2fbc07cd67a38e2f2a5f929a3f11dd3df
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>