linux-marvell: Rework recipe
The recipe has been reworked to not use the Linux Yocto framework as
it, in the recent versions, does not work well with prestine kernel
trees.

The recipe has been renamed for linux-marvell and a new
linux-marvell.inc file has been created to easy the recipe reuse. It
supports config framgments and enables multiple SoCs to share the same
recipe while allow them to specify branches and revisions depending on
the board or SoC family.

Change-Id: I1502dbd7a74610915fd0b29064afe9f672e09ea1
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>